+++ src/governor.c
@@ -9,6 +9,7 @@
 #include <Rinternals.h>
 #include <Rdefines.h>
 
+#include "utils.h"
 
 typedef struct {
   double interval;     // The users specified interval (seconds)
@@ -103,19 +104,6 @@ SEXP gov_init_(SEXP interval_, SEXP alpha_, SEXP alpha_decay_, SEXP alpha_target
 //   long     tv_nsec;       /* nanoseconds */
 // };
 
-
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-// Convert between double and timespec
-//    e.g. 1.5 seconds = timespec(sec = 1, nsec = 500,000,000)
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-static double ts_to_dbl(struct timespec *ts) {
-  return (double)ts->tv_sec + (double)ts->tv_nsec/1e9;
-}
-
-static void dbl_to_ts(double time, struct timespec *ts) {
-  ts->tv_sec  = (time_t)floor(time);
-  ts->tv_nsec = (long)(1e9 * time - floor(time));
-}
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // Use this wait within a for-loop or while-loop