+++ src/utils.h
@@ -1,0 +1,2 @@
+double ts_to_dbl(struct timespec *ts);
+void dbl_to_ts(double time, struct timespec *ts);